make MarkableScrollPane extend SmartScrollPane utilize new utils
